Sad_Rendel/app/requirements.txt
aiopg==1.3.3
anyio==3.3.4
asgiref==3.4.1
async-timeout==4.0.1
click==8.0.3
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2-binary==2.9.2
pycryptodome==3.11.0
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.0
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
